+++ src/IDBObjectStore.js
@@ -38,6 +38,7 @@ IDBObjectStore.__createInstance = function (storeProperties, transaction) {
         me.__autoIncrement = !!storeProperties.autoInc;
 
         me.__indexes = {};
+        me.__indexClones = {};
         me.__indexNames = DOMStringList.__createInstance();
         const indexList = storeProperties.indexList;
         for (const indexName in indexList) {
@@ -58,7 +59,7 @@ IDBObjectStore.__createInstance = function (storeProperties, transaction) {
             },
             set: function (name) {
                 const me = this;
-                if (me.__deleted) {
+                if (me.__deleted || me.__pendingDelete) {
                     throw createDOMException('InvalidStateError', 'This store has been deleted');
                 }
                 IDBTransaction.__assertVersionChange(me.transaction);
@@ -122,6 +123,7 @@ IDBObjectStore.__clone = function (store, transaction) {
  */
 IDBObjectStore.__createObjectStore = function (db, store) {
     // Add the object store to the IDBDatabase
+    store.__pendingCreate = true;
     db.__objectStores[store.name] = store;
     db.objectStoreNames.push(store.name);
 
@@ -141,6 +143,8 @@ IDBObjectStore.__createObjectStore = function (db, store) {
         tx.executeSql(sql, [], function (tx, data) {
             Sca.encode(store.keyPath, function (encodedKeyPath) {
                 tx.executeSql('INSERT INTO __sys__ VALUES (?,?,?,?,?)', [util.escapeSQLiteStatement(store.name), encodedKeyPath, store.autoIncrement, '{}', 1], function () {
+                    delete store.__pendingCreate;
+                    delete store.__deleted;
                     success(store);
                 }, error);
             });
@@ -156,7 +160,7 @@ IDBObjectStore.__createObjectStore = function (db, store) {
  */
 IDBObjectStore.__deleteObjectStore = function (db, store) {
     // Remove the object store from the IDBDatabase
-    store.__deleted = true;
+    store.__pendingDelete = true;
     db.__objectStores[store.name] = undefined;
     db.objectStoreNames.splice(db.objectStoreNames.indexOf(store.name), 1);
 
@@ -164,7 +168,8 @@ IDBObjectStore.__deleteObjectStore = function (db, store) {
     if (storeClone) {
         storeClone.__indexNames = DOMStringList.__createInstance();
         storeClone.__indexes = {};
-        storeClone.__deleted = true;
+        storeClone.__indexClones = {};
+        storeClone.__pendingDelete = true;
     }
 
     // Remove the object store from WebSQL
@@ -181,6 +186,12 @@ IDBObjectStore.__deleteObjectStore = function (db, store) {
             if (data.rows.length > 0) {
                 tx.executeSql('DROP TABLE ' + util.escapeStoreNameForSQL(store.name), [], function () {
                     tx.executeSql('DELETE FROM __sys__ WHERE "name" = ?', [util.escapeSQLiteStatement(store.name)], function () {
+                        delete store.__pendingDelete;
+                        store.__deleted = true;
+                        if (storeClone) {
+                            delete storeClone.__pendingDelete;
+                            storeClone.__deleted = true;
+                        }
                         success();
                     }, error);
                 }, error);
@@ -295,18 +306,33 @@ IDBObjectStore.prototype.__insertData = function (tx, encoded, value, clonedKeyO
     // The `ConstraintError` to occur for `add` upon a duplicate will occur naturally in attempting an insert
     // We process the index information first as it will stored in the same table as the store
     const paramMap = {};
-    const indexPromises = me.indexNames.map((indexName) => {
+    const indexPromises = Object.keys(
+        // We do not iterate `indexNames` as those can be modified synchronously (e.g.,
+        //   `deleteIndex` could, by its synchronous removal from `indexNames`, prevent
+        //   iteration here of an index though per IndexedDB test
+        //   `idbobjectstore_createIndex4-deleteIndex-event_order.js`, `createIndex`
+        //   should be allowed to first fail even in such a case).
+        me.__indexes
+    ).map((indexName) => {
         // While this may sometimes resolve sync and sometimes async, the
         //   idea is to avoid, where possible, unnecessary delays (and
         //   consuming code ought to only see a difference in the browser
         //   where we can't control the transaction timeout anyways).
         return new SyncPromise((resolve, reject) => {
             const index = me.__indexes[indexName];
-            if (index.__pending) {
+            if (
                 // `createIndex` was called synchronously after the current insertion was added to
-                // the transaction queue so not yet ready to be checked (e.g., if two items with
-                /// the same key were added and then a unique index was created, it should not abort
-                // yet, as we're still handling the insertions)
+                //  the transaction queue so although it was added to `__indexes`, it is not yet
+                //  ready to be checked here for the insertion as it will be when running the
+                //  `createIndex` operation (e.g., if two items with the same key were added and
+                //  *then* a unique index was created, it should not continue to err and abort
+                //  yet, as we're still handling the insertions which must be processed (e.g., to
+                //  add duplicates which then cause a unique index to fail))
+                index.__pendingCreate ||
+                // If already deleted (and not just slated for deletion (by `__pendingDelete`
+                //  after this add), we avoid checks
+                index.__deleted
+            ) {
                 resolve();
                 return;
             }
@@ -392,7 +418,7 @@ IDBObjectStore.prototype.add = function (value /* , key */) {
     if (arguments.length === 0) {
         throw new TypeError('No value was specified');
     }
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertActive(me.transaction);
@@ -413,7 +439,7 @@ IDBObjectStore.prototype.put = function (value /*, key */) {
     if (arguments.length === 0) {
         throw new TypeError('No value was specified');
     }
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertActive(me.transaction);
@@ -467,7 +493,7 @@ IDBObjectStore.prototype.__get = function (query, getKey, getAll, count) {
     if (count !== undefined) {
         count = util.enforceRange(count, 'unsigned long');
     }
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertActive(me.transaction);
@@ -569,7 +595,7 @@ IDBObjectStore.prototype['delete'] = function (query) {
         throw new TypeError('A parameter was missing for `IDBObjectStore.delete`.');
     }
 
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertActive(me.transaction);
@@ -601,7 +627,7 @@ IDBObjectStore.prototype.clear = function () {
     if (!(this instanceof IDBObjectStore)) {
         throw new TypeError('Illegal invocation');
     }
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertActive(me.transaction);
@@ -626,7 +652,7 @@ IDBObjectStore.prototype.count = function (/* query */) {
     if (!(me instanceof IDBObjectStore)) {
         throw new TypeError('Illegal invocation');
     }
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertActive(me.transaction);
@@ -641,7 +667,7 @@ IDBObjectStore.prototype.openCursor = function (/* query, direction */) {
     if (!(me instanceof IDBObjectStore)) {
         throw new TypeError('Illegal invocation');
     }
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     const cursor = IDBCursorWithValue.__createInstance(query, direction, me, me, 'key', 'value');
@@ -654,7 +680,7 @@ IDBObjectStore.prototype.openKeyCursor = function (/* query, direction */) {
     if (!(me instanceof IDBObjectStore)) {
         throw new TypeError('Illegal invocation');
     }
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     const [query, direction] = arguments;
@@ -671,7 +697,7 @@ IDBObjectStore.prototype.index = function (indexName) {
     if (arguments.length === 0) {
         throw new TypeError('No index name was specified');
     }
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertNotFinished(me.transaction);
@@ -679,22 +705,14 @@ IDBObjectStore.prototype.index = function (indexName) {
     if (!index || index.__deleted) {
         throw createDOMException('NotFoundError', 'Index "' + indexName + '" does not exist on ' + me.name);
     }
-    /*
-    // const storeClone = me.transaction.objectStore(me.name); // Ensure clone is made if not present
-    // const indexes = storeClone.__indexes;
-    const storeClones = me.transaction.__storeClones;
-    if (!storeClones[me.name] || storeClones[me.name].__deleted) { // The latter condition is to allow store
-                                                         //   recreation to create new clone object
-        storeClones[me.name] = IDBObjectStore.__clone(me, me.transaction);
-    }
 
-    const indexes = storeClones[me.name].__indexes;
-    if (!indexes[indexName]) {
-        indexes[indexName] = IDBIndex.__clone(index, me);
+    if (!me.__indexClones[indexName] ||
+        me.__indexes[indexName].__pendingDelete ||
+        me.__indexes[indexName].__deleted
+    ) {
+        me.__indexClones[indexName] = IDBIndex.__clone(index, me);
     }
-    return indexes[indexName];
-    */
-    return IDBIndex.__clone(index, me);
+    return me.__indexClones[indexName];
 };
 
 /**
@@ -718,11 +736,11 @@ IDBObjectStore.prototype.createIndex = function (indexName, keyPath /* , optiona
         throw new TypeError('No key path was specified');
     }
     IDBTransaction.__assertVersionChange(me.transaction);
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertActive(me.transaction);
-    if (me.__indexes[indexName] && !me.__indexes[indexName].__deleted) {
+    if (me.__indexes[indexName] && !me.__indexes[indexName].__deleted && !me.__indexes[indexName].__pendingDelete) {
         throw createDOMException('ConstraintError', 'Index "' + indexName + '" already exists on ' + me.name);
     }
 
@@ -758,7 +776,7 @@ IDBObjectStore.prototype.deleteIndex = function (name) {
         throw new TypeError('No index name was specified');
     }
     IDBTransaction.__assertVersionChange(me.transaction);
-    if (me.__deleted) {
+    if (me.__deleted || me.__pendingDelete) {
         throw createDOMException('InvalidStateError', 'This store has been deleted');
     }
     IDBTransaction.__assertActive(me.transaction);