+++ tests-mocha/IDBObjectStore/index-spec.js
@@ -23,6 +23,36 @@ describe('IDBObjectStore.index', function () {
         });
     });
 
+    it('should return the same IDBIndex instance within a transaction but not otherwise', function (done) {
+        util.createDatabase('inline', 'inline-index', function (err, db) {
+            if (err) {
+                expect(function () { throw err; }).to.not.throw(Error);
+                done();
+                return;
+            }
+            var tx = db.transaction('inline', 'readwrite');
+            tx.onerror = done;
+
+            var store = tx.objectStore('inline');
+            var index = store.index('inline-index');
+            var index2 = store.index('inline-index');
+            expect(index).to.equal(index2);
+
+            var tx2 = db.transaction('inline', 'readwrite');
+            tx2.onerror = done;
+
+            var store2 = tx2.objectStore('inline');
+            var index3 = store2.index('inline-index');
+
+            expect(index).to.not.equal(index3);
+
+            tx.oncomplete = function () {
+                db.close();
+                done();
+            };
+        });
+    });
+
     it('should have a reference to the object store', function (done) {
         util.createDatabase('inline', 'inline-index', function (err, db) {
             if (err) {